Update README.MD - "Type: For reference" labels link (#5619)
* Create PULL_REQUEST_TEMPLATE.md

Adding first version of PULL_REQUEST_TEMPLATE.md.

The purpose of this template is to improve and simplify writing more accurate Release Notes.

* Update PULL_REQUEST_TEMPLATE.md

Fix typo.

* Update README.MD - "Type: For reference" labels link

Updating a link to issues labeled with "Type: For reference". The old one is not working due to the updates of labels in the past.